requirements.txt
pyodbc==5.1.0
tabulate==0.9.0
openai>=1.54.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-experimental>=0.3.0
langchain-openai>=0.2.0
pandas==2.2.2
ipykernel==6.29.1
python-dotenv>=1.0.0
azure-identity>=1.18.0
azure-ai-inference
gradio
langchain-chroma
anthropic
ipywidgets
jupyter
PyPDF2
langchain-text-splitters>=0.3.0
langchain